v 0.0.4 WIP
Update Core base classes to be compatible with CUDA.
Added core class zDate for date utility methods.

Added Cuda toolset projects , for classes using cuda functionality.
Added solar analysis toolset class. under development.

+++ cpp/headers/zInterface/iterators/zItMeshField.h
@@ -93,7 +93,7 @@ namespace zSpace
 	{
 	protected:
 
-		std::vector<double>::iterator iter;
+		std::vector<float>::iterator iter;
 
 		/*!	\brief pointer to a mesh fieldobject  */
 		zObjMeshScalarField  *fieldObj;
